refactor: Moved `CreateThemePage` component code into Create Theme Page
