+++ src/lib/colors.ts
@@ -1,0 +1,7 @@
+export const COLORS = [
+    "#ffaa40",
+    "#9c40ff",
+    "#ff40aa",
+    "#40ffaa",
+    "#40aaff",
+  ];
